+++ examples/c/hello_world/src/publisher.c
@@ -76,7 +76,7 @@ int main(int argc, char **argv) {
         return 1;
     }
 
-    if (zzdds_register_type_support(dp, "HelloWorld", HelloWorld_compute_key_hash_from_cdr, HelloWorld_get_field_from_cdr) != DDS_RETCODE_OK) {
+    if (HelloWorldTypeSupport_register(dp, "HelloWorld") != DDS_RETCODE_OK) {
         fprintf(stderr, "FAIL: register_type_support failed\n");
         return 1;
     }