+++ examples/zig/shape/shape_main.zig
@@ -1065,6 +1065,7 @@ pub fn main(init: std.process.Init) !void {
     if (!dds.registerTypeSupport(dp, "ShapeType", .{
         .ctx = @ptrCast(&ts_alloc),
         .compute_key_hash = shape_gen.ShapeType.computeKeyHashFromCdr,
+        .compute_key_hash_key_only = shape_gen.ShapeType.computeKeyHashFromCdrKeyOnly,
         .get_field = shape_gen.ShapeType.getFieldFromCdr,
     })) {
         std.log.err("registerTypeSupport() failed", .{});